feat: Implement Firebase backend and secure Gemini proxy
This commit completes the Firebase backend implementation for the Aura project.

- Implements Firestore service functions (`getUserData`, `createUserData`, `getSessionData`, `saveSessionData`, `addMessageToSession`) in `services/firebaseService.ts` for data persistence, following the specified data structure.

- Creates a secure Firebase Cloud Function (`callGemini`) in TypeScript to act as a proxy for the Gemini API. This function handles user authentication and securely manages the API key.

- The frontend service `services/geminiService.ts` is confirmed to use the `callGemini` Cloud Function, ensuring no API keys are exposed on the client-side.

- The Cloud Functions environment is configured with `package.json` and `tsconfig.json` for a TypeScript workflow.
