+++ functions/package.json
@@ -7,20 +7,23 @@
     "shell": "firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
+    "logs": "firebase functions:log",
+    "build": "tsc"
   },
   "engines": {
     "node": "18"
   },
-  "main": "index.js",
+  "main": "lib/index.js",
   "dependencies": {
     "firebase-admin": "^12.0.0",
     "firebase-functions": "^5.0.0",
-    "@google/genai": "^1.6.0"
+    "@google/generative-ai": "^0.11.3"
   },
   "devDependencies": {
     "eslint": "^8.15.0",
-    "eslint-plugin-promise": "^6.0.0"
+    "eslint-plugin-promise": "^6.0.0",
+    "typescript": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0"
   },
   "private": true
 }
